core: Move Glyph to font/glyph.rs
This makes font.rs smaller, as it's a pretty big file.
